rework on product page
