+++ client/src/components/DetailedBookCard.js
@@ -1,15 +1,17 @@
 import React from 'react';
 import { detailedBookCard } from '../assets/css/makeStyles';
-import { Paper, Box, Grid, Chip, Typography } from '@material-ui/core';
+import { Paper, Button, Box, Grid, Chip, Typography } from '@material-ui/core';
+import LendingModal from './LendingModal';
 
 const DetailedBookCard = (props) => {
   const classes = detailedBookCard();
   const { book } = props;
+
   return (
     <Paper className={classes.root}>
       <Grid container justify="flex-start" alignItems="flex-start">
         <Grid item xs={12} sm={4}>
-          <img src={book.img} alt="BookCover" className={classes.cover} />
+          <img src={book.imageUrl} alt="BookCover" className={classes.cover} />
         </Grid>
         <Grid item xs={12} sm={8}>
           <Grid item xs={12}>
@@ -17,18 +19,22 @@ const DetailedBookCard = (props) => {
           </Grid>
           <Grid item xs={12}>
             <Typography component="h3" variant="h4">
-              {' '}
-              {book.title}{' '}
+              {book.title}
             </Typography>
           </Grid>
           <Grid item xs={12}>
-            <Typography variant="h5">{book.autor}</Typography>
+            <Typography variant="h5">{book.author}</Typography>
           </Grid>
           <Grid item xs={12}>
-            <Typography variant="h6">Editora {book.editora}</Typography>
+            <Typography variant="h6">
+              Editora {book.publishingCompany}
+            </Typography>
           </Grid>
           <Grid item xs={12}>
-            <Typography>Ano de publicação: {book.anoPublicacao}</Typography>
+            <Typography>Ano de publicação: {book.publicationYear}</Typography>
+          </Grid>
+          <Grid item xs={12}>
+            <Typography>Edição: {book.edition}</Typography>
           </Grid>
           <Grid item xs={12}>
             <Typography variant="h6">Descrição</Typography>
@@ -46,6 +52,15 @@ const DetailedBookCard = (props) => {
             {book.tag.map((t) => (
               <Chip key={t} label={t} style={{ marginTop: '20px' }} />
             ))}
+          </Grid>
+          <Grid
+            item
+            xs={12}
+            style={{ display: 'flex' }}
+            justify="flex-end"
+            alignContent="flex-end"
+          >
+            <LendingModal />
           </Grid>
         </Grid>
       </Grid>
